Reorganised validator api cache
