+++ validator-api/Cargo.toml
@@ -25,7 +25,6 @@ pin-project = "1.0"
 pretty_env_logger = "0.4"
 rand-07 = { package = "rand", version = "0.7" } # required for compatibility
 rand = "0.8"
-rand_chacha = "0.3.1"
 reqwest = { version = "0.11", features = ["json"] }
 rocket = { version = "0.5.0-rc.1", features = ["json"] }
 serde = "1.0"
